+++ src/design/plone/policy/sensitive.py
@@ -4,6 +4,7 @@ import os
 import re
 import six
 
+
 try:
     from collective.sentry import error_handler
 except ImportError: