+++ src/design/plone/policy/tests/test_initial_structure.py
@@ -1,11 +1,19 @@
 # -*- coding: utf-8 -*-
 from design.plone.policy.testing import DESIGN_PLONE_POLICY_INTEGRATION_TESTING
+from design.plone.policy.utils import TASSONOMIA_AMMINISTRAZIONE
+from design.plone.policy.utils import TASSONOMIA_ARGOMENTI
+from design.plone.policy.utils import TASSONOMIA_NEWS
+from design.plone.policy.utils import TASSONOMIA_ORGANI_GOVERNO
+from design.plone.policy.utils import TASSONOMIA_SERVIZI
 from plone.app.testing import setRoles
 from plone.app.testing import TEST_USER_ID
+from plone.i18n.normalizer.interfaces import IURLNormalizer
+from zope.component import getUtility
 
 import unittest
 
 
+# TODO: rework tests
 class TestInitialStructureCreation(unittest.TestCase):
     layer = DESIGN_PLONE_POLICY_INTEGRATION_TESTING
 
@@ -15,97 +23,88 @@ class TestInitialStructureCreation(unittest.TestCase):
 
         setRoles(self.portal, TEST_USER_ID, ["Manager"])
 
+    def normalize_ids(self, string):
+        return getUtility(IURLNormalizer).normalize(string)
+
+    def check_initial_blocks(self, obj):
+        self.assertEqual(obj.portal_type, "Document")
+        self.assertEqual(len(obj.blocks.values()), 1)
+        self.assertEqual(len(obj.blocks_layout["items"]), 1)
+
+    def check_children_initial_blocks(self, obj):
+        for child in obj.listFolderContents():
+            self.check_initial_blocks(child)
+
     def test_first_level_created(self):
-        self.assertEqual(
-            [x.id for x in self.portal.getFolderContents()],
-            [
-                "amministrazione",
-                "servizi",
-                "novita",
-                "documenti-e-dati",
-                "argomenti",
-            ],
-        )
+        no_blocks = [
+            "prenotazione-appuntamento",
+            "segnalazione-disservizio",
+            "richiesta-di-assistenza",
+            "amministrazione-trasparente",
+            "informativa-privacy",
+            "note-legali",
+            "media-policy",
+        ]
+        for child in self.portal.listFolderContents():
+            if child.id == "leggi-le-faq":
+                self.assertEqual(child.portal_type, "FaqFolder")
+            elif child.id == "dichiarazione-di-accessiblita":
+                self.assertEqual(child.portal_type, "Link")
+            else:
+                self.assertEqual(child.portal_type, "Document")
+                if child.id not in no_blocks:
+                    self.assertEqual(len(child.blocks.values()), 1)
+                    self.assertEqual(len(child.blocks_layout["items"]), 1)
 
     def test_amministrazione_section(self):
         amministrazione = self.portal["amministrazione"]
         self.assertEqual(
             amministrazione.keys(),
-            [
-                "politici",
-                "personale-amministrativo",
-                "organi-di-governo",
-                "aree-amministrative",
-                "uffici",
-                "enti-e-fondazioni",
-                "luoghi",
-            ],
+            [self.normalize_ids(x) for x in TASSONOMIA_AMMINISTRAZIONE],
         )
-        self.assertEqual(amministrazione["politici"].portal_type, "Document")
+        for x in TASSONOMIA_AMMINISTRAZIONE:
+            self.assertEqual(
+                amministrazione.get(self.normalize_ids(x), None).portal_type, "Document"
+            )
+        self.check_children_initial_blocks(amministrazione)
+
         self.assertEqual(
-            amministrazione["personale-amministrativo"].portal_type, "Document"
+            amministrazione["organi-di-governo"].keys(),
+            [self.normalize_ids(x) for x in TASSONOMIA_ORGANI_GOVERNO],
         )
-        self.assertEqual(amministrazione["organi-di-governo"].portal_type, "Document")
-        self.assertEqual(amministrazione["aree-amministrative"].portal_type, "Document")
-        self.assertEqual(amministrazione["uffici"].portal_type, "Document")
-        self.assertEqual(amministrazione["enti-e-fondazioni"].portal_type, "Document")
-        self.assertEqual(amministrazione["luoghi"].portal_type, "Document")
+        for child in amministrazione["organi-di-governo"].listFolderContents():
+            self.assertEqual(child.portal_type, "Document")
+            self.check_initial_blocks(child)
 
     def test_servizi_section(self):
         servizi = self.portal["servizi"]
         self.assertEqual(
             servizi.keys(),
-            [
-                "anagrafe-e-stato-civile",
-                "cultura-e-tempo-libero",
-                "vita-lavorativa",
-                "attivita-produttive-e-commercio",
-                "appalti-pubblici",
-                "catasto-e-urbanistica",
-                "turismo",
-                "mobilita-e-trasporti",
-                "educazione-e-formazione",
-                "giustizia-e-sicurezza-pubblica",
-                "tributi-e-finanze",
-                "ambiente",
-                "salute-benessere-e-assistenza",
-                "autorizzazioni",
-                "agricoltura",
-            ],
+            [self.normalize_ids(x) for x in TASSONOMIA_SERVIZI],
         )
 
         for child in servizi.listFolderContents():
             self.assertEqual(child.portal_type, "Document")
+            self.check_initial_blocks(child)
 
     def test_novita_section(self):
         folder = self.portal["novita"]
-        self.assertEqual(folder.keys(), ["notizie", "comunicati", "eventi"])
-
-        for child in folder.listFolderContents():
-            self.assertEqual(child.portal_type, "Document")
-
-    def test_documenti_e_dati_section(self):
-        folder = self.portal["documenti-e-dati"]
         self.assertEqual(
-            folder.keys(),
-            [
-                "documenti-albo-pretorio",
-                "modulistica",
-                "documenti-funzionamento-interno",
-                "atti-normativi",
-                "accordi-tra-enti",
-                "documenti-attivita-politica",
-                "documenti-tecnici-di-supporto",
-                "istanze",
-                "dataset",
-            ],
+            folder.keys(), [self.normalize_ids(x) for x in TASSONOMIA_NEWS]
         )
 
         for child in folder.listFolderContents():
             self.assertEqual(child.portal_type, "Document")
-            self.assertEqual(len(child.blocks.values()), 1)
-            self.assertEqual(len(child.blocks_layout["items"]), 1)
+            self.check_initial_blocks(child)
 
     def test_argomenti_section(self):
         folder = self.portal["argomenti"]
         self.assertEqual(folder.portal_type, "Document")
+        self.assertEqual(
+            folder.keys(), [self.normalize_ids(x) for x in TASSONOMIA_ARGOMENTI]
+        )
+        self.assertEqual(folder.portal_type, "Document")
+        for child in folder.listFolderContents():
+            self.assertEqual(child.portal_type, "Pagina Argomento")
+            self.assertEqual(len(child.blocks.values()), 1)
+            self.assertEqual(len(child.blocks_layout["items"]), 1)